% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.R
\name{runExample}
\alias{runExample}
\title{runExample}
\usage{
runExample()
}
\arguments{
\item{example}{String with the name of the app to run from inst/ directory.}
}
\value{
shiny app
}
\description{
Run examples from the `inst` directory
}
\examples{
\dontrun{
runExample()
}

}
